[WIP]Problem with  set colors
