+++ openacademy/models/__init__.py
@@ -2,3 +2,4 @@
 
 from . import models
 from . import partner
+from . import session
